# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MOINFO_API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Userapi/GetDepartments/{sectorId}": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "sectorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Userapi/GetMoraqaba/{DepartId}": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "DepartId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Userapi/GetQism": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "MoraqabaId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Userapi/GetDepartNameByIdAsync": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "departId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/Userapi/GetSectorNameByIdAsync": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "sectorId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/Userapi/GetHierarchyNameByIdAsync": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "hierarchyName",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/Userapi/GetSectorAndDepartmentNamesAsync": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "sectorId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "departId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringStringValueTuple"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringStringValueTuple"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringStringValueTuple"
                }
              }
            }
          }
        }
      }
    },
...
    "/Userapi/Create": {
      "post": {
        "tags": [
          "UserApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Userapi/AssignRoles": {
      "post": {
        "tags": [
          "UserApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AspnetUserRoleVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AspnetUserRoleVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AspnetUserRoleVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Userapi/Edit": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "UserApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Userapi/Delete": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/Userapi/GetSpecificUserWithAllPermission": {
      "get": {
        "tags": [
          "UserApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AspnetUserRoleVM": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/MoiEserviceSysUser"
          },
          "role": {
            "$ref": "#/components/schemas/RoleAdmin"
          }
        },
        "additionalProperties": false
      },
...
      "SelectListItem": {
        "type": "object",
        "properties": {
          "disabled": {
            "type": "boolean"
          },
          "group": {
            "$ref": "#/components/schemas/SelectListGroup"
          },
          "selected": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StringStringValueTuple": {
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Please enter JWT with Bearer prefix",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}